Correct uptime check default values
There are some edge cases when using default values, centring mainly
around 0 and how Go interprets this value as equivalent to the
zero-value of an integer. Even when set explicitly to 0 Terraform is not
able to distinguish between the zero-value of an integer and 0.

Additionally a few minimum and maximum values have been set on the
relevant schema attributes.