Improve error messages returned from the provider
For operations that may result in validation errors the API returns an
object in the response. The Terraform provider was ignoring these
validation errors and not indicating to the user what had gone wrong.

This commit improves the error messages returned by the Terraform
provider by creating separate diagnostic values for each violation.