ci: update signing key
When Terraform 1.6 was released the signing key used to verify the
release stopped working, but only for this release. Older releases
continue to install and verify correctly.

In an attempt to resolve the matter we have created a new signing key
and will use this to publish another version of the provider before
escalating the matter to HashiCorp.