chore: support latest go changes
Bump the minimum Go version to 1.21 to support later changes that make
use of later Go features, such as `errors.Join`.

This change also removes the CI component that checked for compatibility
with previous Go versions. This is redundant since the final asset is a
compiled binary under our own control.